Updated CG code to handle longs
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@2352 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ ide/plugins/showtrace/src/main/java/org/overture/ide/plugins/showtrace/viewer/SelectTimeDialog.java
@@ -72,7 +72,7 @@ public class SelectTimeDialog extends TitleAreaDialog
                 if(idx >= 0)
                 {
                     Long theValue = (Long)theTimes.get(idx);
-                    selectedTime = theValue.intValue();
+                    selectedTime = theValue.longValue();
                 } else
                 {
                     selectedTime = currentTime;
